+++ examples/app.py
@@ -8,7 +8,7 @@ def tid(id): return f'todo-{id}'
 @dataclass
 class TodoItem():
     title: str; id: int = -1; done: bool = False
-    def __xt__(self):
+    def __ft__(self):
         show = AX(self.title, f'/todos/{self.id}', id_curr)
         edit = AX('edit',     f'/edit/{self.id}' , id_curr)
         dt = ' (done)' if self.done else ''